Move `BoundingVerticesOfEdge` to dedicated module
